requirements.txt
numpy==1.22.3
Pillow==9.1.0
keras==2.8.0
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
scikit-image==0.19.3
scikit-learn==1.0.2
tensorflow-cpu==2.8.0
gunicorn==20.1.0
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.0.3
